Changed the file-system to use class instances rather than raw objects.
This lead to huge performance improvements as this will cause redux to not use immutability overloading & thus cause a top-level re-render everytime any file was updated (because due to the immutability overloading, they'd also update the root-folder).
To achieve reloads where necessary, I used an event system instead. I simply bubble the events to the top & chose when I want to re-render.

The next step in this process is to remove the `objectPath` variable. It's unreliable when you delete files.
I guess you could update the siblings & all their children whenever you delete a file, but I think it's easier to just pass files as props rather than using a redux selector, and then just re-render via the event system.